\name{acryl}
\alias{acryl}
\docType{data}
\title{NTP raw data - organ weight}
\description{
National Toxicology Program 13 Weeks Gavage Study on Female B6C3F1 Mice administered
with Acrylonitrile (107-13-1, C50215B,5021501)}

\usage{data("acryl")}
\format{
  A data frame with 47 observations on the following 2 variables.
  \describe{
    \item{\code{LungWt}}{a numeric vector}
    \item{\code{dose}}{a factor with levels \code{0} \code{5} \code{10} \code{20} \code{40}}
  }
}

\examples{
data(acryl)
}
\keyword{datasets}
