\name{bivar}
\alias{bivar}
\docType{data}
\title{Variations and malformations in a reproductive study}
\description{Bivariate data}
\usage{data("bivar")}
\format{
  A data frame with 1949 observations on the following 3 variables.
  \describe{
    \item{\code{DAM_ID}}{a numeric vector}
    \item{\code{DEFECT_TYPE}}{a factor with levels \code{} \code{Malformation} \code{Variation}}
    \item{\code{DOSE}}{a numeric vector}
  }
}
\source{National Toxicology Program. Reproductive toxicology. Diethylene glycol dimethyl
ether. Study Number: TER85061 on Swiss CD-1 Mice}
\examples{
data(bivar)
}
\keyword{datasets}
