\name{clin}
\alias{clin}
\docType{data}
\title{Clinical chemistry data: serum triglycerides}
\description{Clinical chemistry data}
\usage{data("clin")}
\format{
  A data frame with 60 observations on the following 19 variables.
  \describe{
    \item{\code{Dose}}{a numeric vector}
    \item{\code{BUN}}{a numeric vector}
    \item{\code{Creat}}{a numeric vector}
    \item{\samp{TP }}{a numeric vector}
    \item{\samp{ALB }}{a numeric vector}
    \item{\code{SerumGlucose}}{a numeric vector}
    \item{\code{CreatKinase}}{a numeric vector}
    \item{\code{ALT}}{a numeric vector}
    \item{\code{SDH}}{a numeric vector}
    \item{\code{ALP}}{a numeric vector}
    \item{\code{Nuc}}{a numeric vector}
    \item{\code{TBA}}{a numeric vector}
    \item{\code{Cholesterol}}{a numeric vector}
    \item{\code{Triglyceride}}{a numeric vector}
    \item{\code{Chloride}}{a numeric vector}
    \item{\code{Sodium}}{a numeric vector}
    \item{\code{Calc}}{a numeric vector}
    \item{\code{PHOS}}{a numeric vector}
    \item{\code{Potassium}}{a numeric vector}
  }
}
\references{Toxicology and Carcinogenesis Studies of Sodium Dichromate Dihydrate (CAS No.
7789-12-0) in F344/N Rats and B6C3F1 Mice (Drinking Water Studies), 2010.}
\examples{
data(clin)
}
\keyword{datasets}
