\name{daphnia}
\alias{daphnia}
\docType{data}
\title{Number of young per adult in daphnia (Ceriodaphnia
dubia) whole effluent toxicity assay}
\description{EPA bioassay}
\usage{data("daphnia")}
\format{
  A data frame with 60 observations on the following 3 variables.
  \describe{
    \item{\samp{Concentration }}{a numeric vector}
    \item{\code{Adults}}{a numeric vector}
    \item{\code{Number_Young}}{a numeric vector}
  }
}
\source{Short-term Methods for Estimating the Chronic Toxicity of Effluents and Receiving
Waters to Freshwater Organisms, Fourth Edition; U.S. Environmental Protection
Agency Office of Water (4303T), Washington, DC 20460,EPA-821-R-02-013; Sec-
tion 13: Test methods Daphnid, Ceriodaphnia Dubia, Survival and Reproduction Test
Methods 1002.0; Table 4}
\examples{
data(daphnia)
}
\keyword{datasets}
