\name{fhema}
\alias{fhema}
\docType{data}
\title{Number of erythrocytes}
\description{Number of erythrocytes in  a 13 weeks study with sodium dichromate dihydrate}
\usage{data("fhema")}
\format{
  A data frame with 58 observations on the following 3 variables.
  \describe{
    \item{\code{dose}}{a numeric vector}
    \item{\code{Ery}}{a numeric vector}
    \item{\code{Dose}}{a factor with levels \code{0} \code{62.5} \code{125} \code{250} \code{500} \code{1000}}
  }
}
\source{National Toxicology Program. 13 Weeks gavage study on female F344 rats admin-
istered with Sodium dichromate dihydrate (VI) (CASRN: 7789-12-0, Study Number:
C20114,TDMS Number:2011402}
\examples{
data(fhema)
}
\keyword{datasets}
