\name{fitz}
\alias{fitz}
\docType{data}
\title{Liver weights of dogs}
\description{Liver weights of dogs in study on mosapride citrate}
\usage{data("fitz")}
\format{
  A data frame with 24 observations on the following 2 variables.
  \describe{
    \item{\code{dose}}{a factor with levels \code{control} \code{low} \code{med} \code{high}}
    \item{\code{Weight}}{a numeric vector}
  }
}
\source{In: T. Yanagawa and Y. Kikuchi. Statistical issues on the determination of the no
observed-adverse-effect levels in toxicology. Environmetrics, 12(4):319–325, June 2001}
\examples{
data(fitz)
}
\keyword{datasets}
