\name{frau}
\alias{frau}
\docType{data}
\title{Time prolongation of reflex response in bevavioral reproductive bioassay}
\description{Data of a bevavioral reproductive bioassay}
\usage{data("frau")}
\format{
  A data frame with 777 observations on the following 7 variables.
  \describe{
    \item{\code{Group}}{a factor with levels \code{1} \code{2} \code{3} \code{4}}
    \item{\code{Litter}}{a factor with levels \code{101} \code{102} \code{103} \code{104} \code{105} \code{106} \code{108} \code{109} \code{110} \code{111} \code{113} \code{115} \code{116} \code{117} \code{119} \code{120} \code{201} \code{202} \code{203} \code{204} \code{205} \code{206} \code{207} \code{208} \code{210} \code{212} \code{213} \code{214} \code{215} \code{216} \code{218} \code{219} \code{220} \code{301} \code{302} \code{303} \code{304} \code{305} \code{306} \code{307} \code{308} \code{310} \code{311} \code{312} \code{313} \code{314} \code{316} \code{317} \code{319} \code{320} \code{401} \code{402} \code{403} \code{404} \code{405} \code{406} \code{407} \code{408} \code{410} \code{412} \code{413} \code{415} \code{416} \code{417} \code{418} \code{419}}
    \item{\code{Sex}}{a factor with levels \code{f} \code{m}}
    \item{\code{Pup}}{a numeric vector}
    \item{\code{Day.1}}{a numeric vector}
    \item{\code{Day.2}}{a numeric vector}
    \item{\code{Day.3}}{a numeric vector}
  }
}
\source{G. Lewin and T Tilmann. Einfluss niederfrequenter elektromagnetischer Felder auf
das sich entwickelnde blutbildende System, das Immunsystem und das ZNS in vivo
- vorhaben 3608s30006 band 1: Hauptbericht. Technical report, Bundesamt für
Strahlenschutz, 2013}
\examples{
data(frau)
}
\keyword{datasets}
