\name{green}
\alias{green}
\docType{data}
\title{Liver basophilia severity scores of second-generation female medaka}
\description{Severity scores}
\usage{data("green")}
\format{
  A data frame with 67 observations on the following 4 variables.
  \describe{
    \item{\code{Dose}}{a numeric vector}
    \item{\code{Severity}}{a numeric vector}
    \item{\code{S12}}{a numeric vector}
    \item{\code{S23}}{a numeric vector}
  }
}
\source{Table 3 of :  J. W. Green, T. A. Springer, A. N. Saulnier, and J. Swintek. Statistical analysis of
histopathological endpoints. Environmental Toxicology and Chemistry, 33(5):1108–
1116, May 2014.}
\examples{
data(green)
}
\keyword{datasets}
