\name{helena1}
\alias{helena1}
\docType{data}
\title{Latency times in Water Morris behavioural test}
\description{Latency times}
\usage{data("helena1")}
\format{
  A data frame with 1152 observations on the following 7 variables.
  \describe{
    \item{\code{group}}{a numeric vector}
    \item{\code{sex}}{a factor with levels \code{F} \code{M}}
    \item{\code{ano}}{a numeric vector}
    \item{\code{day}}{a numeric vector}
    \item{\code{run}}{a numeric vector}
    \item{\code{latency}}{a numeric vector}
    \item{\code{time}}{a numeric vector}
  }
}
\source{C. Faes, M. Aerts, H. Geys, and L. De Schaepdrijver. Modeling spatial learning in
rats based on morris water maze experiments. Pharmaceutical Statistics, 9(1):10–20,
January 2010}
\examples{
data(helena1)
}
\keyword{datasets}
