\name{HgCl}
\alias{HgCl}
\docType{data}
\title{Body weights after 65 and 105 weeks in a longterm
study}}
\description{Body weight data}
\usage{data("HgCl")}
\format{
  A data frame with 180 observations on the following 5 variables.
  \describe{
    \item{\code{Animal}}{a numeric vector}
    \item{\code{Cage}}{a numeric vector}
    \item{\code{Dose}}{a factor with levels \code{2.5 mg/kg} \code{5 mg/kg} \code{Vehicle Control}}
    \item{\code{Week.65}}{a numeric vector}
    \item{\code{Week.105}}{a numeric vector}
  }
}
\source{
National Toxicology Program. Toxicology and Carcinogenesis Studies of
Mercuric Chloride (CAS No. 7487-94-7) in F344 Rats and B6C3F1
Mice (Gavage Studies). Technical Report 408}
\examples{
data(HgCl)
}
\keyword{datasets}
