\name{HIST}
\alias{HIST}
\docType{data}
\title{Historical control proportions of alveolar/bronchial adenomas from 23 bioassays}
\description{Historical control data}
\usage{data("HIST")}
\format{
  A data frame with 23 observations on the following 2 variables.
  \describe{
    \item{\code{tumor}}{a numeric vector}
    \item{\code{rats}}{a numeric vector}
  }
}
\source{R. T. Smythe, D. Krewski, and D. Murdoch. The use of historical control information
in modeling dose-response relationships in carcinogenesis. Statistics & Probability
Letters, 4(2):87–93, 1986}
\examples{
data(HIST)
}
\keyword{datasets}
