\name{hyalin}
\alias{hyalin}
\docType{data}
\title{Hyaline droplets findings}
\description{Incidence of microscopic findings in the kidneys }
\usage{data("hyalin")}
\format{
  A data frame with 48 observations on the following 2 variables.
  \describe{
    \item{\code{dose}}{a factor with levels \code{0} \code{5} \code{10} \code{15} \code{20} \code{30} \code{45} \code{90}}
    \item{\code{droplets}}{a numeric vector}
  }
}
\source{
In Table 1 of: A. Swain, J. Turton, and M. York. Nephrotoxicity of hexachloro-1:3-butadiene in the male Hanover Wistar rat; correlation of minimal histopathological changes with
biomarkers of renal injury. J Appl Toxicol, 32(6):417–428, JUN 2012.}
\examples{
data(hyalin)
}
\keyword{datasets}
