\name{Lirat}
\alias{Lirat}
\docType{data}
\title{Lirat Low-iron rat teratology data}
\description{Reproductive data frim package VGAM}
\usage{data("Lirat")}
\format{
  A data frame with 58 observations on the following 4 variables.
  \describe{
    \item{\code{N}}{a numeric vector}
    \item{\code{R}}{a numeric vector}
    \item{\code{hb}}{a numeric vector}
    \item{\code{grp}}{a numeric vector}
  }
}
\source{Moore, D. F. and Tsiatis, A. (1991) Robust Estimation of the Variance in Moment Methods for Extra-binomial and Extra-Poisson Variation. Biometrics, 47, 383–401}
\references{Shepard, T. H., Mackler, B. and Finch, C. A. (1980) Reproductive studies in the iron-deficient rat. Teratology, 22, 329–334}
\examples{
data(Lirat)
}
\keyword{datasets}
