\name{liv}
\alias{liv}
\docType{data}
\title{Liver weights}
\description{Absolute liver weight data}
\usage{data("liv")}
\format{
  A data frame with 60 observations on the following 3 variables.
  \describe{
    \item{\code{Dose}}{a numeric vector}
    \item{\code{BodyWt}}{a numeric vector}
    \item{\code{LiverWt}}{a numeric vector}
  }
}
\source{National Toxicology Program. 13 Weeks gavage study on female F344 rats admin-
istered with Sodium dichromate}
\examples{
data(liv)
}
\keyword{datasets}
