\name{livMF}
\alias{livMF}
\docType{data}
\title{Relative liver weight data}
\description{Liver weights}
\usage{data("livMF")}
\format{
  A data frame with 120 observations on the following 4 variables.
  \describe{
    \item{\code{Dose}}{a numeric vector}
    \item{\code{BodyWt}}{a numeric vector}
    \item{\code{LiverWt}}{a numeric vector}
    \item{\code{Sex}}{a factor with levels \code{females} \code{males}}
  }
}
\source{National Toxicology Program. 13 Weeks gavage study on female F344 rats admin-
istered with Sodium dichromate dihydrate (VI) (CASRN: 7789-12-0, Study Number:
C20114,TDMS Number:2011402}
\examples{
data(livMF)
}
\keyword{datasets}
