\name{mercur}
\alias{mercur}
\docType{data}
\title{Body weight data of two-years bioassay on mercuric chloride}
\description{Cage-specific body weights}
\usage{data("mercur")}
\format{
  A data frame with 180 observations on the following 9 variables.
  \describe{
    \item{\code{Animal}}{a numeric vector}
    \item{\code{Cage}}{a numeric vector}
    \item{\code{Dose}}{a factor with levels \code{2.5 mg/kg} \code{5 mg/kg} \code{Vehicle Control}}
    \item{\code{Week.1}}{a factor with levels \code{101.9} \code{108.2} \code{110.0} \code{111.8} \code{112.2} \code{113.1} \code{114.4} \code{115.3} \code{116.0} \code{116.1} \code{116.3} \code{117.7} \code{118.8} \code{119.6} \code{119.9} \code{120.3} \code{120.4} \code{120.5} \code{120.8} \code{121.0} \code{121.2} \code{121.9} \code{122.2} \code{122.5} \code{122.7} \code{122.9} \code{123.0} \code{123.1} \code{123.3} \code{123.9} \code{124.1} \code{124.3} \code{124.5} \code{124.6} \code{124.7} \code{124.8} \code{124.9} \code{125.0} \code{125.1} \code{125.2} \code{125.4} \code{125.5} \code{125.6} \code{125.9} \code{126.0} \code{126.4} \code{126.5} \code{126.6} \code{126.8} \code{127.0} \code{127.5} \code{127.7} \code{127.9} \code{128.2} \code{128.4} \code{128.5} \code{128.6} \code{128.8} \code{128.9} \code{129.1} \code{129.2} \code{129.9} \code{130.1} \code{130.2} \code{130.3} \code{130.4} \code{130.5} \code{130.9} \code{131.0} \code{131.2} \code{131.4} \code{131.7} \code{132.0} \code{132.3} \code{132.4} \code{132.8} \code{133.0} \code{133.1} \code{133.3} \code{133.5} \code{134.0} \code{134.1} \code{134.2} \code{134.4} \code{134.6} \code{134.8} \code{134.9} \code{135.0} \code{135.2} \code{135.7} \code{135.8} \code{135.9} \code{136.0} \code{136.1} \code{136.2} \code{136.3} \code{136.9} \code{137.1} \code{137.2} \code{137.3} \code{137.4} \code{137.6} \code{137.8} \code{138.1} \code{138.2} \code{138.7} \code{139.4} \code{139.7} \code{140.2} \code{140.5} \code{140.9} \code{141.0} \code{141.6} \code{141.7} \code{142.6} \code{142.8} \code{143.6} \code{144.0} \code{144.4} \code{144.7} \code{145.1} \code{145.5} \code{146.6} \code{146.8} \code{147.6} \code{147.7} \code{149.9} \code{150.1} \code{150.4} \code{153.4} \code{155.7} \code{180.3}}
    \item{\code{Week.5}}{a factor with levels \code{132.1} \code{143.9} \code{145.7} \code{146.7} \code{147.8} \code{147.9} \code{148.5} \code{151.5} \code{153.6} \code{153.7} \code{154.3} \code{154.6} \code{155.0} \code{155.1} \code{155.2} \code{155.3} \code{156.8} \code{156.9} \code{157.3} \code{157.6} \code{157.7} \code{158.2} \code{158.7} \code{159.7} \code{160.0} \code{160.2} \code{160.3} \code{160.7} \code{160.8} \code{161.5} \code{162.0} \code{162.6} \code{162.7} \code{163.1} \code{163.2} \code{163.4} \code{163.7} \code{163.8} \code{163.9} \code{164.2} \code{164.3} \code{164.9} \code{165.1} \code{165.2} \code{165.5} \code{165.6} \code{165.7} \code{165.8} \code{165.9} \code{166.0} \code{166.1} \code{166.4} \code{166.6} \code{167.1} \code{167.4} \code{167.5} \code{167.6} \code{167.7} \code{167.8} \code{167.9} \code{168.1} \code{168.3} \code{168.5} \code{168.6} \code{168.7} \code{168.9} \code{169.3} \code{169.4} \code{169.6} \code{169.9} \code{170.4} \code{170.5} \code{170.7} \code{170.9} \code{171.1} \code{171.2} \code{171.3} \code{171.5} \code{171.6} \code{171.8} \code{171.9} \code{172.0} \code{172.2} \code{172.3} \code{172.4} \code{172.6} \code{172.7} \code{172.8} \code{172.9} \code{173.0} \code{173.1} \code{173.7} \code{173.9} \code{174.0} \code{174.2} \code{174.3} \code{174.4} \code{174.5} \code{174.6} \code{174.7} \code{174.8} \code{175.1} \code{175.4} \code{175.6} \code{175.7} \code{175.9} \code{176.0} \code{176.2} \code{176.3} \code{176.6} \code{177.1} \code{177.3} \code{177.6} \code{177.8} \code{178.2} \code{178.3} \code{178.4} \code{178.5} \code{179.0} \code{179.1} \code{179.7} \code{179.9} \code{180.0} \code{180.6} \code{181.0} \code{181.2} \code{181.4} \code{181.5} \code{181.6} \code{182.2} \code{182.4} \code{182.8} \code{183.0} \code{183.6} \code{183.9} \code{184.5} \code{184.6} \code{185.6} \code{186.5} \code{187.4} \code{187.8} \code{188.1} \code{188.7} \code{191.9} \code{192.7} \code{193.2} \code{194.7} \code{197.0} \code{197.2} \code{202.2}}
    \item{\code{Week.9}}{a factor with levels \code{149.2} \code{158.3} \code{161.9} \code{166.8} \code{167.0} \code{167.4} \code{170.0} \code{170.1} \code{171.3} \code{172.4} \code{173.3} \code{173.4} \code{175.6} \code{175.7} \code{175.9} \code{176.4} \code{176.5} \code{177.3} \code{177.7} \code{178.1} \code{178.2} \code{178.6} \code{178.7} \code{179.1} \code{179.7} \code{179.9} \code{180.3} \code{180.4} \code{180.8} \code{181.1} \code{182.0} \code{182.1} \code{182.2} \code{182.3} \code{182.6} \code{182.7} \code{183.0} \code{183.3} \code{183.4} \code{184.1} \code{184.4} \code{184.5} \code{184.7} \code{184.9} \code{185.0} \code{185.1} \code{185.4} \code{185.6} \code{185.9} \code{186.0} \code{186.2} \code{186.6} \code{186.7} \code{186.8} \code{186.9} \code{187.1} \code{187.6} \code{188.0} \code{188.4} \code{188.5} \code{188.6} \code{188.9} \code{189.1} \code{189.2} \code{189.3} \code{189.7} \code{189.9} \code{190.0} \code{190.1} \code{190.6} \code{190.7} \code{190.9} \code{191.0} \code{191.5} \code{191.6} \code{191.7} \code{191.8} \code{192.0} \code{192.2} \code{192.9} \code{193.5} \code{193.7} \code{193.8} \code{194.0} \code{194.2} \code{194.4} \code{194.5} \code{194.6} \code{194.7} \code{194.8} \code{194.9} \code{195.0} \code{195.3} \code{195.8} \code{196.0} \code{196.1} \code{196.3} \code{196.5} \code{196.6} \code{197.3} \code{197.4} \code{197.7} \code{198.2} \code{198.4} \code{198.6} \code{198.7} \code{199.1} \code{199.4} \code{199.8} \code{200.1} \code{200.9} \code{201.3} \code{201.5} \code{202.3} \code{202.5} \code{202.9} \code{203.0} \code{204.0} \code{204.1} \code{204.4} \code{204.9} \code{205.6} \code{205.9} \code{206.3} \code{206.8} \code{207.2} \code{207.4} \code{208.4} \code{209.1} \code{210.1} \code{210.5} \code{210.9} \code{211.3} \code{212.2} \code{212.4} \code{212.9} \code{213.5} \code{214.0} \code{214.2} \code{214.4} \code{214.5} \code{215.6} \code{217.1} \code{217.3} \code{218.9} \code{221.4}}
    \item{\code{Week.13}}{a factor with levels \code{160.5} \code{162.3} \code{168.4} \code{169.0} \code{169.5} \code{172.5} \code{176.9} \code{177.4} \code{179.5} \code{180.7} \code{182.5} \code{183.2} \code{184.2} \code{184.5} \code{185.6} \code{186.2} \code{186.4} \code{186.5} \code{187.1} \code{187.4} \code{187.9} \code{188.4} \code{188.6} \code{189.1} \code{189.3} \code{189.4} \code{189.5} \code{190.4} \code{191.0} \code{191.5} \code{191.6} \code{191.7} \code{191.9} \code{192.0} \code{192.3} \code{192.9} \code{193.3} \code{193.5} \code{194.0} \code{194.1} \code{194.2} \code{194.4} \code{194.5} \code{194.6} \code{194.9} \code{195.5} \code{195.6} \code{196.2} \code{196.8} \code{198.0} \code{198.1} \code{198.6} \code{198.7} \code{199.0} \code{199.3} \code{199.7} \code{199.8} \code{200.0} \code{200.4} \code{200.5} \code{200.6} \code{200.9} \code{201.3} \code{201.4} \code{202.2} \code{202.3} \code{202.5} \code{202.6} \code{203.1} \code{203.2} \code{203.5} \code{203.6} \code{203.9} \code{204.1} \code{204.4} \code{205.1} \code{205.3} \code{205.4} \code{205.5} \code{205.8} \code{206.1} \code{206.2} \code{206.4} \code{206.5} \code{206.6} \code{206.9} \code{207.5} \code{207.6} \code{207.7} \code{207.9} \code{208.1} \code{208.3} \code{208.4} \code{208.5} \code{208.6} \code{209.0} \code{209.2} \code{209.4} \code{209.8} \code{209.9} \code{210.1} \code{210.6} \code{210.7} \code{210.9} \code{211.0} \code{211.4} \code{211.6} \code{212.8} \code{213.0} \code{213.2} \code{213.4} \code{213.6} \code{213.7} \code{213.8} \code{214.2} \code{216.0} \code{216.1} \code{216.2} \code{216.3} \code{216.8} \code{217.2} \code{217.3} \code{217.5} \code{217.7} \code{217.8} \code{217.9} \code{218.0} \code{218.3} \code{219.1} \code{219.5} \code{219.6} \code{219.9} \code{220.1} \code{220.3} \code{220.5} \code{220.8} \code{221.2} \code{222.3} \code{222.4} \code{222.8} \code{223.3} \code{223.9} \code{224.2} \code{224.3} \code{226.3} \code{226.8} \code{227.2} \code{228.2} \code{228.7} \code{229.8} \code{231.7} \code{239.1} \code{239.4} \code{245.2}}
    \item{\code{Week.53}}{a factor with levels \code{200.6} \code{208.3} \code{211.7} \code{217.3} \code{222.8} \code{223.4} \code{224.1} \code{227.6} \code{228.2} \code{228.9} \code{229.4} \code{230.1} \code{230.3} \code{230.4} \code{230.5} \code{230.9} \code{231.8} \code{232.4} \code{232.7} \code{233.2} \code{233.9} \code{235.3} \code{235.6} \code{235.8} \code{235.9} \code{237.1} \code{237.3} \code{237.6} \code{238.0} \code{238.4} \code{238.7} \code{239.5} \code{240.0} \code{240.5} \code{240.6} \code{241.2} \code{241.4} \code{241.9} \code{242.3} \code{242.5} \code{242.8} \code{243.4} \code{243.5} \code{243.6} \code{244.0} \code{244.3} \code{244.5} \code{244.7} \code{244.8} \code{245.3} \code{246.1} \code{246.2} \code{246.9} \code{247.0} \code{247.7} \code{247.8} \code{247.9} \code{248.1} \code{249.7} \code{250.6} \code{250.9} \code{251.1} \code{251.3} \code{251.5} \code{252.0} \code{252.4} \code{252.8} \code{253.1} \code{253.3} \code{254.5} \code{254.9} \code{255.3} \code{255.7} \code{256.5} \code{257.0} \code{257.3} \code{257.7} \code{257.8} \code{258.0} \code{258.2} \code{258.7} \code{259.2} \code{259.3} \code{259.6} \code{260.2} \code{260.4} \code{260.6} \code{260.8} \code{261.2} \code{261.3} \code{261.5} \code{262.5} \code{263.0} \code{263.8} \code{264.8} \code{264.9} \code{265.1} \code{266.2} \code{267.4} \code{267.5} \code{268.2} \code{268.3} \code{268.8} \code{270.3} \code{271.2} \code{271.3} \code{271.6} \code{271.7} \code{272.0} \code{272.2} \code{272.7} \code{273.1} \code{273.6} \code{273.8} \code{274.1} \code{275.3} \code{275.4} \code{276.5} \code{277.0} \code{277.8} \code{278.2} \code{279.5} \code{280.0} \code{280.3} \code{280.8} \code{281.1} \code{281.4} \code{282.5} \code{283.0} \code{283.1} \code{284.5} \code{286.2} \code{286.4} \code{287.4} \code{287.7} \code{288.5} \code{288.6} \code{289.1} \code{291.1} \code{291.6} \code{292.1} \code{292.9} \code{293.6} \code{296.6} \code{297.0} \code{297.7} \code{300.2} \code{300.3} \code{301.1} \code{306.0} \code{318.5}}
    \item{\code{Week.105}}{a factor with levels \code{177.1} \code{184.6} \code{184.9} \code{227.6} \code{251.3} \code{254.9} \code{257.7} \code{258.3} \code{258.4} \code{260.1} \code{267.0} \code{271.7} \code{272.0} \code{273.6} \code{278.6} \code{279.5} \code{279.9} \code{280.0} \code{287.9} \code{290.1} \code{290.5} \code{291.7} \code{293.7} \code{296.0} \code{297.9} \code{298.1} \code{298.5} \code{299.1} \code{300.6} \code{300.7} \code{301.8} \code{302.4} \code{302.9} \code{303.3} \code{303.5} \code{303.8} \code{304.4} \code{306.4} \code{306.7} \code{307.3} \code{308.0} \code{309.6} \code{310.5} \code{313.6} \code{314.6} \code{315.4} \code{318.1} \code{318.2} \code{319.0} \code{321.6} \code{323.1} \code{324.7} \code{324.9} \code{326.2} \code{326.5} \code{326.9} \code{327.2} \code{327.3} \code{327.9} \code{328.1} \code{329.5} \code{330.1} \code{330.8} \code{330.9} \code{332.3} \code{332.6} \code{334.5} \code{334.8} \code{335.3} \code{338.0} \code{338.2} \code{339.1} \code{343.9} \code{345.6} \code{345.7} \code{349.3} \code{352.5} \code{359.0} \code{360.6} \code{361.5} \code{364.8} \code{369.2} \code{377.6}}
  }
}
\source{
National Toxicology Program. 2 years bioassay of mercuric chloride on female rats.
Body weight data. 1993 (study number C60173)}
\examples{
data(mercur)
}
\keyword{datasets}
