\name{mm}
\alias{mm}
\docType{data}
\title{Clinical chemistry data:  blood urea nitrogen}
\description{BUN data of the 13 weeks study}
\usage{data("mm")}
\format{
  A data frame with 60 observations on the following 2 variables.
  \describe{
    \item{\code{BUN}}{a numeric vector}
    \item{\code{Dose}}{a factor with levels \code{0} \code{62.5} \code{125} \code{250} \code{500} \code{1000}}
  }
}
\source{Toxicology and Carcinogenesis Studies of Sodium Dichromate Dihydrat}
\examples{
data(mm)
}
\keyword{datasets}
