\name{mn}
\alias{mn}
\docType{data}
\title{Micronucleus assay  data}
\description{MN counts}
\usage{data("mn")}
\format{
  A data frame with 25 observations on the following 3 variables.
  \describe{
    \item{\code{group}}{a factor with levels \code{Control} \code{D188} \code{D375} \code{D750} \code{Positive}}
    \item{\code{animal}}{a numeric vector}
    \item{\code{MN}}{a numeric vector}
  }
}
\source{G. Engelhardt. In vivo micronucleus test in mice with 1-phenylethanol. Archives of Toxicology, 80(12):868–872, December 2006}
\examples{
data(mn)
}
\keyword{datasets}
