\name{paul}
\alias{paul}
\docType{data}
\title{Litter-specific number of abnormal pups}
\description{Per-litter analysis data}
\usage{data("paul")}
\format{
  A data frame with 568 observations on the following 3 variables.
  \describe{
    \item{\code{Treatment}}{a factor with levels \code{Control} \code{High} \code{Low} \code{Medium}}
    \item{\code{Litter}}{a numeric vector}
    \item{\code{Abnormal}}{a numeric vector}
  }
}
\source{S. R. Paul. Analysis of proportions of affected fetuses in teratological experiments.Biometrics, 38(2):361–370, 1982}
\examples{
data(paul)
}
\keyword{datasets}
