\name{statorgans}
\alias{statorgans}
\docType{data}
\title{Body and organ weights of multiple sites}
\description{Body, Heart, Liver, Lung, Spleen and Thymus Weight}
\usage{data("statorgans")}
\format{
  A data frame with 60 observations on the following 7 variables.
  \describe{
    \item{\code{dose}}{a numeric vector}
    \item{\code{BodyWt}}{a numeric vector}
    \item{\code{Heart}}{a numeric vector}
    \item{\code{Liver}}{a numeric vector}
    \item{\code{Lung}}{a numeric vector}
    \item{\code{Spleen}}{a numeric vector}
    \item{\code{Thymus}}{a numeric vector}
  }
}
\source{National Toxicology Program. 13 Weeks gavage study on female F344 rats admin-
istered with Sodium dichromate}
\examples{
data(statorgans)

}
\keyword{datasets}
