\name{tr580}
\alias{tr580}
\docType{data}
\title{NTP raw data - metaplasia of olfactory epithelium}
\description{National Toxicology Program. National Toxicology Program. Carcinogenicity Studies
on beta-Picoline in F344 rats and B6C3F1 MICE. TR-580, 2014}
\usage{data("tr580")}
\format{
  A data frame with 400 observations on the following 3 variables.
  \describe{
    \item{\code{Sex}}{a factor with levels \code{female} \code{male}}
    \item{\code{Dose}}{a factor with levels \code{Control} \code{High} \code{Low} \code{Medium}}
    \item{\code{Tumor}}{a numeric vector}
  }
}
\examples{
data(tr580)
}
\keyword{datasets}
