\name{tubepi}
\alias{tubepi}
\docType{data}
\title{Histopathological tubular epithelia findings}

\description{Histopathological tubular epithelia findings
histopathological in the NTP
P-Cresidine carcinogenicity study 
}
\usage{data("tubepi")}
\format{
  A data frame with 40 observations on the following 5 variables.
  \describe{
    \item{\code{X}}{a numeric vector}
    \item{\code{Anino}}{a numeric vector}
    \item{\code{Group}}{a factor with levels \code{Control} \code{High} \code{Low} \code{Medium}}
    \item{\code{TubularEpithelia}}{a numeric vector}
    \item{\code{Without}}{a numeric vector}
  }
}
\source{National Cancer Institute. Carcinogenesis technical report series no. 142 1979 bioassay
of p-cresidine for possible carcinogenicity nci-cg-tr-142. Technical report, 1979.
}
\examples{
data(tubepi)
}
\keyword{datasets}
