\name{wuleon}
\alias{wuleon}
\docType{data}
\title{Pup weights and malformation rates}
\description{Multiple endpoints in reproductive bioassay: the pup weights and malformation rates}
\usage{data("wuleon")}
\format{
  A data frame with 1028 observations on the following 4 variables. b 
  \describe{
    \item{\code{Litter}}{a numeric vector}
    \item{\code{Dose}}{a numeric vector}
    \item{\code{Weight}}{a numeric vector}
    \item{\code{Malformation}}{a numeric vector}
  }
}
\source{B. L. Wu and A. R. de Leon. Gaussian copula mixed models for clustered mixed
outcomes, with application in developmental toxicology. Journal of Agricultural Bio-
logical and Environmental Statistics, 19(1):39–56, March 2014}
\examples{
data(wuleon)
}
\keyword{datasets}
